Move LowerSpatialFourVelocity out of Xcts
