+++ src/Elliptic/Systems/Xcts/HydroQuantities.hpp
@@ -52,28 +52,4 @@ struct HydroQuantitiesCompute : ::Tags::Variables<HydroTags>, db::ComputeTag {
   }
 };
 
-/*!
- * \brief Computes $u_i=W \gamma_{ij} v^j$, where $W$ is the Lorentz factor,
- * $\gamma_{ij}$ is the spatial metric, and $v^j$ is the spatial velocity.
- *
- * This compute item is intended for observations in a pure XCTS solve where the
- * hydro quantities are retrieved directly from the background solution/data.
- *
- * \see HydroQuantitiesCompute
- */
-struct LowerSpatialFourVelocityCompute
-    : hydro::Tags::LowerSpatialFourVelocity<DataVector, 3, Frame::Inertial>,
-      db::ComputeTag {
-  using base =
-      hydro::Tags::LowerSpatialFourVelocity<DataVector, 3, Frame::Inertial>;
-  using argument_tags =
-      tmpl::list<hydro::Tags::SpatialVelocity<DataVector, 3, Frame::Inertial>,
-                 gr::Tags::SpatialMetric<DataVector, 3>,
-                 hydro::Tags::LorentzFactor<DataVector>>;
-  static void function(const gsl::not_null<tnsr::i<DataVector, 3>*> result,
-                       const tnsr::I<DataVector, 3>& spatial_velocity,
-                       const tnsr::ii<DataVector, 3>& spatial_metric,
-                       const Scalar<DataVector>& lorentz_factor);
-};
-
 }  // namespace Xcts::Tags
